dashboards/requirements.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.1.4
snowflake-connector-python[pandas]